Avoid duplicate code in datamanagerderive/importmodal by adding callback in the inputchecker
