+++ packages/algomancy-gui/src/algomancy_gui/data_page/datamanagementderivemodal.py
@@ -180,55 +180,6 @@ def derive_data_callback(n_clicks, selected_data_key, derived_name, invalid_deri
         )
     except Exception as e:
         return no_update, "", False, f"Problem with deriving: {str(e)}", True, False
-""" 
-@callback(
-    [
-        Output(DM_IMPORT_MODAL_NAME_INPUT, "invalid"),
-        Output(DM_IMPORT_MODAL_FEEDBACK, "children"),
-    ],
-    Input(DM_IMPORT_MODAL_NAME_INPUT, "value"),
-    State(ACTIVE_SESSION, "data")
-)
-def dataset_name_invalid(value, session_id: str):
-    
-        Input field for dataset name gets a red border, and a red error message appears below the field.
-
-        Checks the input and displays feedback if one of the scenarios below holds:
-        1) input contains characters that are not alphanumeric, hyphens or underscores
-        2) input is already in use for another saved dataset
-
-        Args:
-            value: String containing user input for dataset name
-            session_id: ID of the active session
-
-        Returns:
-            tuple: (invalid, feedback_children) where:
-            - invalid: Boolean indicating if feedback should be shown
-            - feedback_children: String containing feedback message
-    
-    # No dataset_name defined yet
-    if not value:
-        return False, ""
-
-    # Dataset_name not character safe
-    is_invalid = not bool(re.fullmatch(r"[A-Za-z0-9_-]+", value))
-
-    if is_invalid:
-        feedback_msg = "This is not a valid dataset name. Please only use alphanumeric characters, hyphens and underscores."
-        return is_invalid, feedback_msg
-
-    # Dataset_name already exists
-    sm: ScenarioManager = get_scenario_manager(get_app().server, session_id)
-    dataset_names = sm.get_data_keys()
-    is_invalid = value in dataset_names
-
-    if is_invalid:
-        feedback_msg = "This is not a valid dataset name. A dataset with this name already exists."
-        return is_invalid, feedback_msg
-
-    # Valid Dataset_name
-    return False, ""
-"""
 
 @callback(
     Output(DM_DERIVE_MODAL, "is_open"),
@@ -257,51 +208,4 @@ def toggle_modal_derive(open_clicks, close_clicks, is_open):
         return not is_open
     return is_open
 
-@callback(
-    [
-        Output(DM_DERIVE_SET_NAME_INPUT, "invalid"),
-        Output(DM_DERIVE_MODAL_FEEDBACK, "children"),
-        Output(DM_DERIVE_MODAL_SUBMIT_BTN, "disabled"),
-        Output(DM_DERIVE_MODAL_SUBMIT_BTN, "color"), #todo: css styling
-    ],
-    Input(DM_DERIVE_SET_NAME_INPUT, "value"),
-    State(ACTIVE_SESSION, "data")
-)
-def dataset_name_invalid(value, session_id: str):
-    """
-        In case of an invalid dataset name, the input field for dataset name gets a red border, and a red error message appears below the field
-        This also disables the use of the Import button.
-
-        Checks dataset_name validity via the below three scenarios:
-        1) user input is empty
-        2) user input contains characters that are not alphanumeric, hyphens or underscores
-        3) user input is already in use for another saved dataset
-        Feedback is displayed and the import button is made inactive, until none of the scenarios hold.
-
-        Args:
-            value: String containing user input for dataset name
-            session_id: ID of the active session
-
-        Returns:
-            tuple: (invalid, feedback_children) where:
-            - invalid: Boolean indicating whether feedback will be shown
-            - feedback_children: String containing feedback message
-            - disabled: Boolean indicating whether the import button will be disabled
-            - color: String describing the color of the import button (green if enabled, gray if disabled)
-    """
-    # No dataset_name defined yet
-    if not value:
-        return False, "", True, "secondary"
-
-    # Dataset_name not character safe
-    if not InputChecker.is_character_safe(value):
-        feedback_msg = "This is not a valid dataset name. Please only use alphanumeric characters, hyphens and underscores."
-        return True, feedback_msg, True, "secondary"
-
-    # Dataset_name already exists
-    if InputChecker.name_exists(value, session_id):
-        feedback_msg = "This is not a valid dataset name. A dataset with this name already exists."
-        return True, feedback_msg, True, "secondary"
-
-    # Valid Dataset_name
-    return False, "", False, "primary"
+InputChecker.register_name_validator_static(DM_DERIVE_SET_NAME_INPUT, DM_DERIVE_MODAL_FEEDBACK, DM_DERIVE_MODAL_SUBMIT_BTN, ACTIVE_SESSION)
